fix crash on report phase when result size bigger than 2GB

+++ src/result_summarizer.h
@@ -12,6 +12,7 @@
 #include <string>
 #include <sstream>
 #include <boost/thread.hpp>
+#include <stdint.h>
 
 #include "statistics.h"
 
@@ -31,7 +32,7 @@ class ResultSummarizer{
 		MPIResource::DatabaseInfo database_info;
 		std::vector<MPI::Request> reqs;
 		std::vector<char *> result_data_list;
-		std::vector<int> result_size_list;
+		std::vector<uint64_t> result_size_list;
 		AlignerCommon::AligningCommonParameters parameters;
 		std::ostream *os;
 		boost::mutex *check_request_mutex;
@@ -44,32 +45,32 @@ class ResultSummarizer{
 	
 	int LoadResultFile(std::vector<std::vector<Result> > &results_list,
 					   AlignmentTask task);
-	int LoadResultFile(char **ptr,int *size,AlignmentTask task);
-	void SendResult(char *data,int size,AlignmentTask task,int target_rank);
+	int LoadResultFile(char **ptr,uint64_t *size,AlignmentTask task);
+	void SendResult(char *data,uint64_t size,AlignmentTask task,int target_rank);
 	void RecvResult(std::vector<std::vector<Result> > &results_list,
 					AlignmentTask &task);
 
 
 	void SerializeResult(std::vector<std::vector<Result> > &results_list,
-						 char **ptr,int *size);
+						 char **ptr,uint64_t *size);
 	void DeserializeResult(std::vector<std::vector<Result> > &results_list,
-						   char *ptr,int size);
+						   char *ptr,uint64_t size);
 
 	void GatherResultMaster(int query_chunk_size,int mpi_size,
 							AligningParameters &parameters, DatabaseInfo &database_info);
 	void GatherResultWorker(int rank,
 							AligningParameters &parameters, DatabaseInfo &database_info);
    
-	void AddList(AlignmentTask task,int size);
+	void AddList(AlignmentTask task,uint64_t size);
 		
 	void BuildResultTargetMap(int query_chunk_size, int mpi_size);
  private:
 	std::vector<AlignmentTask > task_list; // searched results file on each process
-	std::vector<int> size_list;            // file size of results file on each process
+	std::vector<uint64_t> size_list;            // file size of results file on each process
 	std::string tmp_dirname_;
 	std::string output_filename_;
 	std::vector<int> result_target_map;   // map of query block to report phase node number
- 	std::vector<std::vector<int> > result_size_map; // sizeof (query,db) 's results file
+ 	std::vector<std::vector<uint64_t> > result_size_map; // sizeof (query,db) 's results file
 	
 	
 	void DeleteResultFile(AlignmentTask task);
